First working of a stepper test program
The stepper motor is moving, but not correctly. It was necessary to
implement a replacement for the Arduino micros() function, as the
one from the Longan Nano after a short while returns only -1.

The motor is not moving correctly, and there's a timing issue in
the Stepper class. It should not be necessary to add a delay to
the Stepper::step() function.